+++ src/util/get-delimiter-position.js
@@ -7,6 +7,7 @@ const beginDisplayMath = ["$$", "\\[","\\("];
 const endDisplayMath = ["$$", "\\]","\\)"];
 const beginInlineMath = ["$", "\\("];
 const endInlineMath = ["$", "\\)"];
+// TODO: Add "$$" to Inline math delimiter array
 
 
 function MatchIndex(matchStr, matchIndex) {